+++ HISTORY.md
@@ -1,5 +1,10 @@
 # HISTORY
 
+## Unreleased
+
+* Improve default layout and actually remove Erb layout file
+* Prevent load order issue with i18n-spec matchers
+
 ## 0.3.0
 
 * Properly installed stationed executable [ariejan]