+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Improve default layout and actually remove Erb layout file
 * Prevent load order issue with i18n-spec matchers
 * Allow passing additional interpolation arguments to record-based page titles
+* Add extra action-default record page title translation key
 
 ## 0.3.0