+++ HISTORY.md
@@ -1,6 +1,6 @@
 # HISTORY
 
-## Unreleased
+## 0.4.0
 
 * Improve default layout and actually remove Erb layout file
 * Prevent load order issue with i18n-spec matchers